fix memory leak: cache isFirstRun, close response bodies inline
isFirstRun called GetCompletedTaskRuns (SELECT * with no LIMIT) every
task cycle, loading all historical rows into memory. At 50K rows this
allocated ~48 MB per call. Replace with HasCompletedTaskRun (EXISTS
query) and cache the result on platformRunner so it's checked at most
once per task.

Also fix defer resp.Body.Close() inside pagination loops in Shopify
and Squarespace scrapers — close bodies inline after each page.

+++ internal/database/models.go
@@ -89,7 +89,7 @@ type TaskRun struct {
 	CompletedAt    sql.NullTime   `json:"completed_at"`
 	Status         string         `json:"status"`
 	ErrorMessage   sql.NullString `json:"error_message"`
-	StatusCode     sql.NullInt32  `json:"status_code"`
+	StatusCode     sql.NullInt16  `json:"status_code"`
 	ResponseTimeMs sql.NullInt32  `json:"response_time_ms"`
 	CacheStatus    sql.NullString `json:"cache_status"`
 }
